move method stuff to methods
