+++ Libraries/Opc.Ua.Client/Methods/EventMethod.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
-using Opc.Ua.Client.Methods;
 
-namespace Opc.Ua.Client.Events
+namespace Opc.Ua.Client.Methods
 {
     /// <summary>
     /// A standard event method
